Clarify the simulation information of OVPsim and riscvOVPsimPlus
OVPsim is developed and maintained by Imperas. The core simulation
platform is proprietary software; it is available free of charge for
non-commercial usage. Various processor (RISC-V included), peripheral
and platform models are available as free software under the Apache
License version 2.0.

This patch removed the duplicated entries and distinguished beteween
OVPsim and riscvOVPsimPlus.